Give links a visual cue
Make links default to being underlined, and then opt-out of this for
several blocks, that already look like a link or would be reasonably
expected to act like a link.

This improves accessibility as the user no longer has to rely upon
colour to identify a link within the 'main content' of the page.